+++ api/src/main/resources/db/changelog/changes/create-table-teams.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS teams
     country varchar(255) not null,
     city varchar(255) not null,
     balance decimal(19,2) not null,
-    commission double not null,
+    commission decimal(19,2) not null,
     CONSTRAINT teams_pk PRIMARY KEY (id)
     );
 --rollback DROP TABLE teams;
